backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
psycopg2-binary==2.9.10
redis==5.2.1
django-redis==5.4.0
python-dotenv==1.0.1
sentry-sdk==2.19.2
Pillow==11.0.0
gunicorn==23.0.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
anthropic==0.40.0
openai==1.58.1
httpx==0.27.2
google-generativeai==0.8.3
requests==2.32.3
